pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "teams-agent"
version = "0.1.0"
description = "CLI daemon that monitors Microsoft Teams messages, auto-replies using GPT, and alerts via Telegram"
requires-python = ">=3.11"
dependencies = [
    "msal>=1.28.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "typer[all]>=0.12.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
teams-agent = "teams_agent.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
